requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.0
drf-yasg2==1.19.4
Fiona==1.9.0
geopandas==0.12.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
munch==2.5.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
